测试

+++ src/main/java/com/github/ompc/robot/hexapod/dragoon/component/ServoCtlPiCom.java
@@ -23,7 +23,7 @@ public interface ServoCtlPiCom extends PiCom {
         private final int index;
 
         // 目标角度
-        private final int angle;
+        private final float angle;
 
         /**
          * 构造舵机命令
@@ -31,7 +31,7 @@ public interface ServoCtlPiCom extends PiCom {
          * @param index 舵机编号
          * @param angle 旋转目标角度
          */
-        public ServoCmd(int index, int angle) {
+        public ServoCmd(int index, float angle) {
             this.index = index;
             this.angle = angle;
         }
@@ -50,7 +50,7 @@ public interface ServoCtlPiCom extends PiCom {
          *
          * @return 旋转目标角度
          */
-        public int getAngle() {
+        public float getAngle() {
             return angle;
         }
     }
